-Take out components not being used -Change DevTeam -Make Join SHPE button go to the form straight -Change Readme to Jaysprogram
