+++ src/app/_components/HeroSection.tsx
@@ -31,7 +31,7 @@ export default function HeroSection() {
         {/* CTA Button */}
         <div className="flex justify-center">
           <a
-            href="/signUp"
+            href="https://form.jotform.com/70387424224151"
             target="_blank"
             rel="noopener noreferrer"
             className="inline-flex items-center justify-center
