--- src/app/_components/Join.tsx
@@ -1,10 +1,0 @@
-"use client";
-import React from "react";
-
-export default function Join() {
-  return (
-    <div className="px-4 sm:px-10 py-14 max-w-5xl mx-auto">
-   
-    </div>
-  );
-}
