--- src/app/alumni/AlumniInfo.tsx
@@ -1,28 +1,0 @@
-import { type Member } from "~/app/_components/MemberCard";
-
-
-
-
-export const alumniList: Member[] = [
-    {
-        name: "Hernan Hernandez-Garcia",
-        role: "PRESIDENT",
-        picture: "/alumni/hernan.jpg",
-        linkedin: "https://www.linkedin.com/in/hernanhernan/",
-    },
-    {
-        name: "Vimalys Rivera Caceres",
-        role: "EXTERNAL VICE PRESIDENT",
-        picture: "/alumni/vimalys.jpg",
-        linkedin: "https://www.linkedin.com/in/vimalys-rivera-caceres/",
-    },
-    {
-        name: "Ariana Rodriguez Velez",
-        role: "INTERNAL VICE PRESIDENT",
-        picture: "/alumni/ariana.jpg",
-        linkedin: "https://www.linkedin.com/in/arodriguezvelez/",
-    },
-    //to add more alumni later
-
-];
-
