--- src/app/alumni/_components/constants.ts
@@ -1,1 +1,0 @@
-export const DEFAULT_PAGE_SIZE = 20 as const;
