+++ src/app/board/_components/ContactCard.tsx
@@ -1,7 +1,7 @@
 import Flag from "react-world-flags";
 import iso from "iso-3166-1";
 import { BsEnvelope, BsLinkedin } from "react-icons/bs";
-import { type Member } from "./MemberCard";
+import { type Member } from "../../_components/MemberCard";
 import Image from "next/image";
 
 type ContactCardProps = {
