+++ src/app/page.tsx
@@ -16,12 +16,15 @@ export default function HomePage() {
       {/* full page */}
       <div className="bg-[#F7F8F9] text-black overflow-x-hidden flex flex-col">
         <Navbar />
+        
         <HeroSection />   
+
         <AboutSection />
+
         <WeeklyEvents />
 
         <Slideshow />
-  
+
       </div>
     </>
   );
